requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.1
langchain-tavily>=0.0.3 # for TavilySearch
langchain-exa>=0.1.0 # for Exa Search
# langchain-brave_search>=0.0.3 # for Brave Search (if not included in langchain-community)
duckduckgo-search>=5.0.0
# 预装主流提供商以支持动态配置 (以下为可选安装，根据config.yaml配置)
langchain-groq>=0.1.0
langchain-fireworks>=0.1.0
langchain-mistralai>=0.1.0
streamlit>=1.30.0
python-dotenv>=1.0.0
PyYAML>=6.0
tavily-python>=0.3.0
requests>=2.30.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
ddgs>=0.1.0 # for DuckDuckGo Search
PySide6>=6.0.0
PyInstaller>=5.0
semantic-chunking>=0.1.1
streamlit-agraph>=0.0.45
networkx>=3.0
leidenalg>=0.10.0
python-igraph>=0.10.0
fpdf2>=2.7.0
EbookLib>=0.18
markdown>=3.4.0
EbookLib>=0.17.1
SQLAlchemy>=2.0.0